+++ common/screens/src/commonMain/kotlin/app.campfire.common.screens/Screens.kt
@@ -101,6 +101,7 @@ data class SettingsScreen(
     Root,
     Account,
     Appearance,
+    Downloads,
     Playback,
     Sleep,
     About,