+++ features/libraries/ui/src/iosMain/kotlin/app/campfire/libraries/ui/detail/permission/NotificationPermission.ios.kt
@@ -1,0 +1,8 @@
+package app.campfire.libraries.ui.detail.permission
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun rememberPostNotificationPermissionState(onPermissionResult: (Boolean) -> Unit): PermissionState {
+  return PermissionState.Granted
+}